Use default inspect formatting, remove omitempty, update docs
