+++ cli/command/trust/inspect.go
@@ -2,12 +2,11 @@ package trust
 
 import (
 	"encoding/json"
-	"fmt"
 	"sort"
 
 	"github.com/docker/cli/cli"
 	"github.com/docker/cli/cli/command"
-	"github.com/pkg/errors"
+	"github.com/docker/cli/cli/command/inspect"
 	"github.com/spf13/cobra"
 	"github.com/theupdateframework/notary/tuf/data"
 )
@@ -18,36 +17,24 @@ func newInspectCommand(dockerCli command.Cli) *cobra.Command {
 		Short: "Return low-level information about keys and signatures",
 		Args:  cli.RequiresMinArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return inspectTrustInfo(dockerCli, args)
+			return runInspect(dockerCli, args)
 		},
 	}
 	return cmd
 }
 
-func inspectTrustInfo(cli command.Cli, remotes []string) error {
-	trustRepoInfoList := []trustRepo{}
-	for _, remote := range remotes {
-		trustInfo, err := getRepoTrustInfo(cli, remote)
-		if err != nil {
-			return err
-		}
-		if trustInfo == nil {
-			continue
-		}
-		trustRepoInfoList = append(trustRepoInfoList, *trustInfo)
+func runInspect(dockerCli command.Cli, remotes []string) error {
+	getRefFunc := func(ref string) (interface{}, []byte, error) {
+		i, err := getRepoTrustInfo(dockerCli, ref)
+		return nil, i, err
 	}
-	trustInspectJSON, err := json.Marshal(trustRepoInfoList)
-	if err != nil {
-		return errors.Wrap(err, "error while serializing trusted repository info")
-	}
-	fmt.Fprintf(cli.Out(), string(trustInspectJSON))
-	return nil
+	return inspect.Inspect(dockerCli.Out(), remotes, "", getRefFunc)
 }
 
-func getRepoTrustInfo(cli command.Cli, remote string) (*trustRepo, error) {
+func getRepoTrustInfo(cli command.Cli, remote string) ([]byte, error) {
 	signatureRows, adminRolesWithSigs, delegationRoles, err := lookupTrustInfo(cli, remote)
 	if err != nil {
-		return nil, err
+		return []byte{}, err
 	}
 	// process the signatures to include repo admin if signed by the base targets role
 	for idx, sig := range signatureRows {
@@ -61,37 +48,36 @@ func getRepoTrustInfo(cli command.Cli, remote string) (*trustRepo, error) {
 	signerRoleToKeyIDs := getDelegationRoleToKeyMap(delegationRoles)
 
 	for signerName, signerKeys := range signerRoleToKeyIDs {
-		signerList = append(signerList, trustSigner{signerName, signerKeys})
+		signerKeyList := []trustKey{}
+		for _, keyID := range signerKeys {
+			signerKeyList = append(signerKeyList, trustKey{ID: keyID})
+		}
+		signerList = append(signerList, trustSigner{signerName, signerKeyList})
 	}
 	sort.Slice(signerList, func(i, j int) bool { return signerList[i].Name > signerList[j].Name })
 
 	for _, adminRole := range adminRolesWithSigs {
 		switch adminRole.Name {
 		case data.CanonicalRootRole:
-			adminList = append(adminList, trustSigner{"Root", adminRole.KeyIDs})
+			rootKeys := []trustKey{}
+			for _, keyID := range adminRole.KeyIDs {
+				rootKeys = append(rootKeys, trustKey{ID: keyID})
+			}
+			adminList = append(adminList, trustSigner{"Root", rootKeys})
 		case data.CanonicalTargetsRole:
-			adminList = append(adminList, trustSigner{"Repository", adminRole.KeyIDs})
+			targetKeys := []trustKey{}
+			for _, keyID := range adminRole.KeyIDs {
+				targetKeys = append(targetKeys, trustKey{ID: keyID})
+			}
+			adminList = append(adminList, trustSigner{"Repository", targetKeys})
 		}
 	}
 	sort.Slice(adminList, func(i, j int) bool { return adminList[i].Name > adminList[j].Name })
 
-	return &trustRepo{
+	return json.Marshal(trustRepo{
+		Name:              remote,
 		SignedTags:        signatureRows,
 		Signers:           signerList,
 		AdminstrativeKeys: adminList,
-	}, nil
-}
-
-// trustRepo represents consumable information about a trusted repository
-type trustRepo struct {
-	SignedTags        trustTagRowList `json:",omitempty"`
-	Signers           []trustSigner   `json:",omitempty"`
-	AdminstrativeKeys []trustSigner   `json:",omitempty"`
-}
-
-// trustSigner represents a trusted signer in a trusted repository
-// a signer is defined by a name and list of key IDs
-type trustSigner struct {
-	Name string   `json:",omitempty"`
-	Keys []string `json:",omitempty"`
+	})
 }
