+++ cli/command/trust/testdata/trust-inspect-uninitialized.golden
@@ -1,0 +1,1 @@
+[]
